pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "tele-ai-userbot"
version = "0.1.0"
description = "Telegram userbot for automatic AI translation."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "aiosqlite>=0.21.0",
  "cachetools>=5.5.2",
  "lingua-language-detector>=2.1.1",
  "openai>=2.0.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.1",
  "python-telegram-bot>=22.0",
  "telethon>=1.41.2",
  "tenacity>=9.1.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
  "pytest-asyncio>=1.1.0",
]

[project.scripts]
tele-ai = "tele_ai.main:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/tele_ai"]
